==> code/05-GSEA-analysis_210501.ipynb <==
# !pip install gseapy

%matplotlib inline
%config InlineBackend.figure_format='retina' # mac
%load_ext autoreload
%autoreload 2
import pandas as pd
import gseapy as gp
import matplotlib.pyplot as plt
import os

exp= "/Users/ltao/Projects/mecom/exp/210210/bulkrna/"

[i for i in os.listdir(exp) if ('norm' in i and 'txt' in i)]

# dataset='norm_GeneMat.txt' #aml
# dataset='norm_lthsc_aavs_mecom_GeneMat.txt' # lthsc
dataset='norm_lthsc_res_mecom_GeneMat.txt'
# dataset='norm_lthsc_aavs_res_GeneMat.txt'

# method="log2_ratio_of_classes"
method="t_test"
# method='signal_to_noise'
# method='ratio_of_classes'

output=exp+'subset/'+dataset[:-4]+"_"+method
aml_norm=pd.read_csv(exp+dataset, sep='\t')
x=[g.split('_')[1] for g in aml_norm.index.to_list()]
aml_norm['Name']=x
# aml_norm['DESCRIPTION']='aml_aavs_vs_mecom'
aml_norm['DESCRIPTION']='aml_mecom_vs_aavs'

gsea_exp=aml_norm[['Name','DESCRIPTION']+aml_norm.columns[:-2].to_list()].reset_index(drop=True)
# gsea_exp=aml_norm[['Name','DESCRIPTION']+aml_norm.columns[::-1][2:].to_list()].reset_index(drop=True)

# enrichr libraries are supported by gsea module. Just provide the name

gs_res = gp.gsea(data=gsea_exp, # or data='./P53_resampling_data.txt'
                 gene_sets='/Users/ltao/Projects/mecom/exp/210210/mecom.gmt', # enrichr library names
                 cls= exp+"lthsc_res_mecom.cls", # cls=class_vector
#                  cls= exp+"lthsc.cls", # cls=class_vector

#                  cls= exp+"aml.cls", # cls=class_vector

                 # set permutation_type to phenotype if samples >=15
                 permutation_type='gene_set',
                 permutation_num=1000, # reduce number to speed up test
                 outdir=output,  # do not write output to disk
                 no_plot=False, # Skip plotting
                 method=method,
                 max_size=1000,
                 processes=4,
                 seed= 7,
                 format='pdf')


# tcag

"TCGA gene exp stratified by MECOM exp.csv"

dataset="TCGA gene exp stratified by MECOM exp.csv"
# method="log2_ratio_of_classes"
method="diff_of_classes"
# method='ratio_of_classes'
output=exp+'subset/'+dataset[:-4]+"_"+method

aml_norm=pd.read_csv(exp+dataset)
aml_norm['Name']=aml_norm['attrib_name']
aml_norm['DESCRIPTION']='tcga'
gsea_exp=aml_norm[['Name','DESCRIPTION','Avg MECOM low','Avg MECOM high']].reset_index(drop=True)
gsea_exp

dataset="TCGA gene exp stratified by MECOM exp.csv"
# method="log2_ratio_of_classes"
method="diff_of_classes"
# method='ratio_of_classes'
output=exp+'subset/'+dataset[:-4]+"_"+method

aml_norm=pd.read_csv(exp+dataset)
aml_norm['Name']=aml_norm['attrib_name']
aml_norm['DESCRIPTION']='tcga'
# gsea_exp=aml_norm[['Name','DESCRIPTION']+aml_norm.columns[1:3].to_list()].reset_index(drop=True)
gsea_exp=aml_norm[['Name','DESCRIPTION','Avg MECOM low','Avg MECOM high']].reset_index(drop=True)

# enrichr libraries are supported by gsea module. Just provide the name

gs_res = gp.gsea(data=gsea_exp, # or data='./P53_resampling_data.txt'
                 gene_sets='/Users/ltao/Projects/mecom/exp/210210/mecom.gmt', # enrichr library names
                 cls= exp+"tcga.cls", # cls=class_vector
#                  cls= exp+"aml.cls", # cls=class_vector

                 # set permutation_type to phenotype if samples >=15
                 permutation_type='gene_set',
                 permutation_num=1000, # reduce number to speed up test
                 outdir=output,  # do not write output to disk
                 no_plot=False, # Skip plotting
                 method=method,
                 max_size=1000,
#                  method='t_test',
                 
#                  method='log2_ratio_of_classes',
                 processes=4, seed= 7,
                 format='pdf')


## TCGA 24

output

dataset="TCGA_24.csv"
# method="log2_ratio_of_classes"
method="diff_of_classes"
# method='ratio_of_classes'
output=exp+'subset/'+dataset[:-4]+"_"+method

aml_norm=pd.read_csv(exp+dataset)
aml_norm['Name']=aml_norm['attrib_name']
aml_norm['DESCRIPTION']='tcga'
gsea_exp=aml_norm[['Name','DESCRIPTION','average surv>24 months','average surv<24 months']].reset_index(drop=True)
gsea_exp

# enrichr libraries are supported by gsea module. Just provide the name

gs_res = gp.gsea(data=gsea_exp, # or data='./P53_resampling_data.txt'
                 gene_sets='/Users/ltao/Projects/mecom/exp/210210/mecom.gmt', # enrichr library names
                 cls= exp+"tcga24.cls", # cls=class_vector
#                  cls= exp+"aml.cls", # cls=class_vector

                 # set permutation_type to phenotype if samples >=15
                 permutation_type='gene_set',
                 permutation_num=1000, # reduce number to speed up test
                 outdir=output,  # do not write output to disk
                 no_plot=False, # Skip plotting
                 method=method,
                 max_size=1000,
#                  method='t_test',
                 
#                  method='log2_ratio_of_classes',
                 processes=4, seed= 7,
                 format='pdf')


# psudobulk 10x

ls psdobulk_10x.csv

dataset="psdobulk_10x.csv"
# method="log2_ratio_of_classes"
method="diff_of_classes"
# method='ratio_of_classes'
output=exp+'subset/'+dataset[:-4]+"_"+method+'_ma'

aml_norm=pd.read_csv(dataset)
aml_norm['Name']=aml_norm['genes']
aml_norm['DESCRIPTION']='10x'
# gsea_exp=aml_norm[['Name','DESCRIPTION','aavs','mecom']].reset_index(drop=True)
gsea_exp=aml_norm[['Name','DESCRIPTION','mecom','aavs']].reset_index(drop=True)

gsea_exp


gs_res = gp.gsea(data=gsea_exp, # or data='./P53_resampling_data.txt'
                 gene_sets='/Users/ltao/Projects/mecom/exp/210210/mecom.gmt', # enrichr library names
#                  cls= exp+"tcga24.cls", # cls=class_vector
                 cls= exp+"10x_ma.cls", # cls=class_vector

                 # set permutation_type to phenotype if samples >=15
                 permutation_type='gene_set',
                 permutation_num=1000, # reduce number to speed up test
                 outdir=output,  # do not write output to disk
                 no_plot=False, # Skip plotting
                 method=method,
                 max_size=1000,
#                  method='t_test',
                 
#                  method='log2_ratio_of_classes',
                 processes=4, seed= 7,
                 format='pdf')


# non_lthsc 10x psudobulk

dataset="nonlthsc_10x_1500.csv"
# method="log2_ratio_of_classes"
method="diff_of_classes"
# method='ratio_of_classes'
output=exp+'subset/'+dataset[:-4]+"_"+method+'_ma'

aml_norm=pd.read_csv(dataset)
aml_norm['Name']=aml_norm['genes']
aml_norm['DESCRIPTION']='10x'
# gsea_exp=aml_norm[['Name','DESCRIPTION','aavs','mecom']].reset_index(drop=True)
gsea_exp=aml_norm[['Name','DESCRIPTION','mecom','aavs']].reset_index(drop=True)

gsea_exp


gs_res = gp.gsea(data=gsea_exp, # or data='./P53_resampling_data.txt'
                 gene_sets='/Users/ltao/Projects/mecom/exp/210210/mecom.gmt', # enrichr library names
#                  cls= exp+"tcga24.cls", # cls=class_vector
                 cls= exp+"10x_ma.cls", # cls=class_vector

                 # set permutation_type to phenotype if samples >=15
                 permutation_type='gene_set',
                 permutation_num=1000, # reduce number to speed up test
                 outdir=output,  # do not write output to disk
                 no_plot=False, # Skip plotting
                 method=method,
                 max_size=1000,
#                  method='t_test',
                 
#                  method='log2_ratio_of_classes',
                 processes=4, seed= 7,
                 format='pdf')


#pre rank
import os
from gseapy.plot import gseaplot, heatmap


# aml

ls bulkrna/M4*

de="CCLE all AML high vs low DEG.xlsx"
aml_de=pd.read_excel(exp+de, sheet_name="depmap_download-2")
aml_de
aml_pre=aml_de[['Gene','EffectSize']].reset_index(drop=True)
aml_pre=pd.DataFrame.from_records(aml_pre.values)
aml_pre=aml_pre[aml_pre[1]>=1]
pre_res = gp.prerank(rnk=aml_pre, gene_sets= '/Users/ltao/Projects/mecom/exp/210210/mecom.gmt',
                 processes=4,
                 permutation_num=1000, # reduce number to speed up testing
                     
                 outdir=exp+de[:-4]+"_"+"pre_rank", format='png', seed=6)
de_pre[de]=pre_res
# Make sure that ``ofname`` is not None, if you want to save your figure to disk

print ('>>>'*11,de,"<<<"*11)

print(pre_res.res2d)

aml_pre

# M4 prerank

de_pre={}

de="M4 AML hi vs low expression.rnk.txt"
aml_de=pd.read_csv(exp+de, sep='\t')
aml_de
aml_pre=aml_de[['Feature','Effect_size']].reset_index(drop=True)
aml_pre=pd.DataFrame.from_records(aml_pre.values)
pre_res = gp.prerank(rnk=aml_pre, gene_sets= '/Users/ltao/Projects/mecom/exp/210210/mecom.gmt',
                 processes=4,
                 permutation_num=1000, # reduce number to speed up testing
                     
                 outdir=exp+de[:-4]+"_"+"pre_rank", format='pdf', seed=6)
de_pre[de]=pre_res
# Make sure that ``ofname`` is not None, if you want to save your figure to disk

print ('>>>'*11,de,"<<<"*11)

print(pre_res.res2d)
